requirements.txt
pytest
selenium
psycopg2-binary
allure-pytest
requests
